requirements/production.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/production.in
#
asgiref==3.7.2
    # via django
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
cffi==1.15.1
    # via cryptography
cryptography==41.0.4
    # via django-rest-knox
django==4.2.5
    # via
    #   -r requirements/production.in
    #   django-appconf
    #   django-debug-toolbar
    #   django-defender
    #   django-extensions
    #   django-factory-boy
    #   django-health-check
    #   django-jazzmin
    #   django-rest-knox
    #   djangorestframework
    #   drf-spectacular
django-appconf==1.0.5
    # via django-imagekit
django-debug-toolbar==4.2.0
    # via -r requirements/production.in
django-defender==0.9.7
    # via -r requirements/production.in
django-extensions==3.2.3
    # via -r requirements/production.in
django-factory-boy==1.0.0
    # via -r requirements/production.in
django-health-check==3.17.0
    # via -r requirements/production.in
django-imagekit==4.1.0
    # via -r requirements/production.in
django-jazzmin==2.6.0
    # via -r requirements/production.in
django-migrations-formatter==1.0.0
    # via -r requirements/production.in
django-probes==1.7.0
    # via -r requirements/production.in
django-rest-framework==0.1.0
    # via -r requirements/production.in
django-rest-knox==4.2.0
    # via -r requirements/production.in
djangorestframework==3.14.0
    # via
    #   django-rest-framework
    #   django-rest-knox
    #   drf-spectacular
drf-spectacular==0.26.4
    # via -r requirements/production.in
factory-boy==3.3.0
    # via django-factory-boy
faker==19.6.2
    # via factory-boy
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.19.1
    # via drf-spectacular
jsonschema-specifications==2023.7.1
    # via jsonschema
markupsafe==2.1.3
    # via werkzeug
pilkit==2.0
    # via django-imagekit
pillow==10.0.1
    # via -r requirements/production.in
psycopg2==2.9.7
    # via -r requirements/production.in
pycparser==2.21
    # via cffi
python-dateutil==2.8.2
    # via faker
python-dotenv==1.0.0
    # via -r requirements/production.in
pytz==2023.3.post1
    # via djangorestframework
pyyaml==6.0.1
    # via drf-spectacular
redis==5.0.0
    # via django-defender
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.10.3
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via
    #   django-imagekit
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
uritemplate==4.1.1
    # via drf-spectacular
werkzeug==2.3.7
    # via -r requirements/production.in
